Refactor: move LearnerState to learner_state.rs
